requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
autograd=1.5=pypi_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.9.14=py38h06a4308_0
charset-normalizer=2.1.1=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=pypi_0
cycler=0.11.0=pypi_0
deap=1.3.3=pypi_0
docopt=0.6.2=pypi_0
fonttools=4.37.3=pypi_0
future=0.18.3=pypi_0
idna=3.4=pypi_0
imageio=2.22.0=pypi_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.11.2=pypi_0
matplotlib=3.6.0=pypi_0
ncurses=6.3=h5eee18b_3
networkx=2.8.6=pypi_0
numpy=1.23.3=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.5.2=pypi_0
pillow=9.2.0=pypi_0
pip=22.1.2=py38h06a4308_0
pipreqs=0.4.13=pypi_0
pretty-errors=1.2.25=pypi_0
ptflops=0.7.2.2=pypi_0
pyhocon=0.3.60=pypi_0
pymop=0.2.4=pypi_0
pyparsing=3.0.9=pypi_0
pyrapl=0.2.3.1=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch-ignite=0.4.11=pypi_0
pytorchcv=0.0.67=pypi_0
pytz=2022.6=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.1=pypi_0
seaborn=0.12.1=pypi_0
setuptools=63.4.1=py38h06a4308_0
six=1.16.0=pypi_0
sklearn=0.0=pypi_0
sqlite=3.39.2=h5082296_0
tabulate=0.9.0=pypi_0
termcolor=2.2.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.8.12=pypi_0
timm=0.4.9=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.11.0+cu113=pypi_0
torchaudio=0.11.0+cu113=pypi_0
torchmetrics=1.4.0=pypi_0
torchstat=0.0.7=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.12.0+cu113=pypi_0
tqdm=4.64.1=pypi_0
typing-extensions=4.3.0=pypi_0
urllib3=1.26.12=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.15.0=pypi_0
xz=5.2.5=h7f8727e_1
yarg=0.1.9=pypi_0
zlib=1.2.12=h5eee18b_3
